user_google_login/requirements.txt
google-cloud-firestore==2.2.0
google-cloud-secret-manager==2.7.0
google-auth==2.0.0
defusedxml==0.7.1
protobuf==3.20.0